Move EditProductModal components to product page
